Fix SiteMap.fetchPackages
