+++ Sources/App/Core/Search.swift
@@ -216,23 +216,6 @@ enum Search {
 }
 
 
-private extension SQLSelectBuilder {
-    // sas 2020-06-05: workaround `direction: SQLExpression` signature in SQLKit
-    // (should be SQLDirection)
-    func orderBy(_ expression: SQLExpression, _ direction: SQLDirection = .ascending) -> Self {
-        return self.orderBy(SQLOrderBy(expression: expression, direction: direction))
-    }
-
-    func column(_ matchType: Search.MatchType) -> Self {
-        column(matchType.sqlAlias)
-    }
-
-    func column(_ expression: SQLExpression, as alias: SQLExpression) -> Self {
-        column(SQLAlias(expression, as: alias))
-    }
-}
-
-
 public struct SQLOrderByGroup: SQLExpression {
     public let orderByClauses: [SQLOrderBy]
